Rename ArticleEntity to Article
