+++ api/entities/article_entity.go
@@ -1,6 +1,6 @@
-package entity
+package entities
 
-type ArticleEntity struct {
+type Article struct {
 	ID string `json:"id"`
 	Title string `json:"title"`
 	URL string `json:"url"`
