+++ api/repository/repository.go
@@ -3,10 +3,10 @@ package repository
 import (
 	"context"
 
-	"github.com/kentaro-m/spider/api/entity"
+	"github.com/kentaro-m/spider/api/entities"
 )
 
 type ArticleRepository interface {
-	Get(ctx context.Context) ([]*entity.ArticleEntity, error)
-	Create(ctx context.Context, a *entity.ArticleEntity) (error)
+	Get(ctx context.Context) ([]*entities.Article, error)
+	Create(ctx context.Context, a *entities.Article) (error)
 }
